Changed sum request type to post.

+++ client/model_sum_request.go
@@ -10,6 +10,6 @@
 package openapi
 // SumRequest struct for SumRequest
 type SumRequest struct {
-	A int64 `json:"a,omitempty"`
-	B int64 `json:"b,omitempty"`
+	A int32 `json:"a,omitempty"`
+	B int32 `json:"b,omitempty"`
 }
